Cleaning a bit after messy refactoring.
