+++ src/MiNET/MiNET/Net/Reliability.cs
@@ -1,0 +1,15 @@
+namespace MiNET.Net
+{
+	public enum Reliability
+	{
+		UNRELIABLE = 0,
+		UNRELIABLE_SEQUENCED = 1,
+		RELIABLE = 2,
+		RELIABLE_ORDERED = 3,
+		RELIABLE_SEQUENCED = 4,
+		UNRELIABLE_WITH_ACK_RECEIPT = 5,
+
+		RELIABLE_WITH_ACK_RECEIPT = 6,
+		RELIABLE_ORDERED_WITH_ACK_RECEIPT = 7
+	}
+}
